[#154111360] Architect wants platform-wide refactoring to ensure logical modules of Basis (Chain-Link fabrication, Choice, and Arrangement), Evaluation (ingestion of content from Library/Patterns/Instruments), and Cache (implementing Caffeine), reducing overall  complexity in order to bring our milestones within reach.  \   * [core/evaluation] DigestHash evaluation deprecates any legacy LibraryHash implementation   |   * [core/basis]   |   - Moved to its own package, out of core/work/basis   |   - implements Evaluation for reading entities bound to a given chain   |   - instantiates separate libraryEvaluation for tertiary-bound entities to chain (anything in library of something bound to chain)   |   * [cache/entity] EntityCache deprecates any individual caching of readOne()   |   * [cache/evaluation]   |   - EvaluationFactory is accessed only behind the EvaluationProvider which caches any given evaluation for N seconds.   |   - Implement Caffeine or manual time expiry for all caches   |   * [core/cache] CacheKey.of(....) computes unique signature for any unique collection of entities, or access to entities.
[#153110625] Logout, expect redirect to logged-out home view
 \
  * [hub] for a permit-all route:
      - if an access token was provided, we're going to treat this as a user auth
      - Required, for example, to implement an idempotent /logout endpoint that redirects somewhere, never returning a 401, whether or not the user is auth'd

+++ hub/src/main/java/io/xj/hub/resource/auth/nullify/AuthNullifyResource.java
@@ -4,13 +4,14 @@ package io.xj.hub.resource.auth.nullify;
 import io.xj.core.CoreModule;
 import io.xj.core.access.AccessControlProvider;
 import io.xj.core.access.impl.Access;
+import io.xj.core.dao.UserDAO;
 import io.xj.core.model.user_role.UserRoleType;
 import io.xj.core.transport.HttpResponseProvider;
-import io.xj.core.dao.UserDAO;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 
+import javax.annotation.security.PermitAll;
 import javax.annotation.security.RolesAllowed;
 import javax.jws.WebResult;
 import javax.ws.rs.GET;
@@ -19,6 +20,7 @@ import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
 import java.io.IOException;
+import java.util.Objects;
 
 /**
  Nullify current authentication
@@ -37,15 +39,19 @@ public class AuthNullifyResource {
    */
   @GET
   @WebResult
-  @RolesAllowed(UserRoleType.USER)
-  public Response getCurrentAuthentication(@Context ContainerRequestContext crc) throws IOException {
+  @PermitAll
+  public Response nullifyAuthentication(@Context ContainerRequestContext crc) throws IOException {
     Access access = Access.fromContext(crc);
-    try {
+
+    if (Objects.nonNull(access)) try {
       userDAO.destroyAllTokens(access.getUserId());
       return response.internalRedirectWithCookie("", accessControlProvider.newExpiredCookie());
 
     } catch (Exception e) {
       return response.failure(e);
     }
+
+    else return response.internalRedirectWithCookie("", accessControlProvider.newExpiredCookie());
   }
+
 }
